+++ CHANGELOG.md
@@ -41,6 +41,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (bug fixes) #fse-142 | apps/mission 1.0.1 | Change amount of decimals displayed: 6 instead of 2
 - (bug fixes) #fse-142 | apps/staking 1.0.1 | Change amount of decimals displayed: 6 instead of 2
 - (bug fixes) #fse-142 | apps/staking 1.0.1 | Display tooltip on topBarStaking with 6 decimals amount if amount is bigger than 0
+- (bug fixes) #fse-142 | packages/helpers 1.0.1 | Create function for display TopBar Tooltip
+
 - (chore) #fse-142 | evmos-wallet 1.0.2 | Export EVMOS_DECIMALS
 
 ## 1.0.0 - 2023-04-28